backend/requirements.txt
# Django Backend Requirements
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.4
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
Pillow==10.1.0

# API Documentation
drf-spectacular==0.27.0

# Additional utilities
setuptools==69.0.0
